fix: censor sensitive log messages
When working with secrets and there's an error when applying/patching
the resource, we end up printing the full resource. The full resource
may contain secret material.

To prevent logging secret material we first check if we're about to add
a secret, if we are and we errror out then print a censored error
message.

Signed-off-by: Bheesham Persaud <me@bheesham.com>