Add optional built-in substitution of last attempted revision
The purpose of this new feature (which is CLI-conditional, and defaults
to false in order to preserve compatibility) is to allow to refer to
the revision that is currently being applied in the resource manifests,
via the standard substitutions mechanism.

By enabling the feature, the controller will always add the
FLUX_ARTIFACT_REVISION key to the substitutions map, with the value of
the artifact revision as provided by the source controller.

If, however, the artifact revision matches the following pattern:

    ([^@]+)@sha1:(.+)

then it will also add the following two keys:

- FLUX_ARTIFACT_REF: the 1st group in the pattern, representing the
  branch or ref part for source-control based revisions (e.g. Git)
- FLUX_ARTIFACT_SHA: the 2nd group in the pattern, representing the SHA
  usually in source-control based revisions (e.g. Git).

Signed-off-by: arikkfir <arik@kfirs.com>